remove buncha fwd decls, fix makefile, new file for common shit even looks alright in some specific files
