+++ SAUCE/gamestates/state_demo.h
@@ -2,8 +2,7 @@
 CODE_BANK_PUSH("XCD_BANK_02")
 
 void mouse_and_cursor();
-
-
+void set_settings();
 
 const unsigned char instructions[448]={
 0x04,0xfe,0x04,0x3f,0xff,0x04,0x07,0x20,0x21,0x22,0x23,0x24,0x25,0x26,0x27,0x28,
