+++ SAUCE/include_menus.h
@@ -2,6 +2,9 @@
 CODE_BANK_PUSH("XCD_BANK_03")
 
 #include "defines/menunametable.h"
+
+#include "menustates/common_xcd_bank_03.c"
+
 #include "menustates/levelselection.c"
 #include "menustates/progressbar.c"
 #include "menustates/customize_screen.c"
