+++ SAUCE/menustates/titlescreen.c
@@ -4,8 +4,6 @@ CODE_BANK_PUSH("XCD_BANK_03")
 #if __VS_SYSTEM
 void color_picker();
 #endif
-void check_if_music_stopped();
-void movement();
 void bounds_check();
 
 void title_cube_shit();
@@ -17,17 +15,9 @@ void title_ball_shit();
 void title_swing_shit();
 void title_ship_shit();
 
-void loop_routine_update();
-void dec_mouse_timer();
 void roll_new_mode();
 void settings();
 void set_title_icon();
-void state_demo();
-void mouse_and_cursor();
-void colorinc();
-void colordec();
-void set_settings();
-extern uint8_t famistudio_song_speed;
 
 const uint8_t xbgmlookuptable[];
 
@@ -1190,23 +1180,6 @@ void roll_new_mode() {
 	set_title_icon();
 }			
 
-void dec_mouse_timer() {
-	kandoframecnt++;
-	if (kandoframecnt & 1 && mouse_timer) mouse_timer--;	
-}		
-
-void loop_routine_update() {
-	newrand();
-	ppu_wait_nmi();
-	#if __VS_SYSTEM
-	check_if_music_stopped();
-	#endif
-	music_update();
-	oam_clear();
-	mouse_and_cursor();
-}		
-
-
 void bounds_check() {
 	if (currplayer_y_small >= 160) {
 		currplayer_y_small = 160;
@@ -1347,16 +1320,6 @@ void title_robot_shit() {
 	}
 					
 }
-
-void check_if_music_stopped() {
-	#if __VS_SYSTEM
-		tmp2 = (newrand() & 31) + 1;
-		if (famistudio_song_speed == 0x80) { music_play(xbgmlookuptable[tmp2]); }
-	#else
-		if (songplaying && famistudio_song_speed == 0x80) { music_play(xbgmlookuptable[song]); }
-		else if (famistudio_song_speed == 0x80) { music_play(xbgmlookuptable[menu_music]); }
-	#endif
-}	
 
 #if __VS_SYSTEM
 void color_picker() {
